% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastregls_main_function.R
\name{fastregls}
\alias{fastregls}
\title{Run the FastRegLS algorithm to fit the parameters of the MELS model}
\usage{
fastregls(y, X, Z, id, nIter = 1, getSigmaSE = TRUE)
}
\arguments{
\item{y}{Response variable. Should be a vector.}

\item{X}{Mean model fixed-effects design matrix. First column should be an intercept.}

\item{Z}{Within-subject variance model fixed-effects design matrix. First column should be an intercept.}

\item{id}{Vector of subject IDs. Should be in order from 1 to the number of subjects.}

\item{nIter}{Number of times to execute the inner loop. Defaults to 1.}

\item{getSigmaSE}{Whether or not to return the standard errors of the random location and random scale variances. Defaults to \code{TRUE}. If \code{FALSE}, returns NA in these fields.}
}
\value{
A list with the following elements:
\item{beta}{Estimate of beta, the mean fixed-effect coefficients}
\item{tau}{Estimate of tau, the WS variance fixed-effect coefficients}
\item{tau_l}{Estiamte of tau_l, the effect of the random location effect on the WS variance}
\item{sigma_sq_nu}{Estimate of random location variance}
\item{sigma_sq_omega}{Estimate of random scale variance}
\item{se_beta}{Standard error of beta}
\item{se_tau}{Standard error of tau}
\item{se_tau_l}{Standard error of tau_l}
\item{se_sigma_sq_nu}{Standard error of random location variance}
\item{se_sigma_sq_omega}{Standard error of random scale variance}
\item{nu}{Posterior means of the random location effects, ordered by subject}
\item{omega}{Posterior means of the random scale effects, ordered by subject}
\item{marginal_variances}{Diagonal of the fitted covariance matrix}
}
\description{
Some text
}
